TASK: Provide one HttpFactory that implements all interfaces of PSR 17
Injecting multiple objects to create PSR Http objects has been tedious in the past.

To overcome this the implementations of the PSR 17 methods are extracted into traits
and a central HttpFactory that implements all those interfaces is created.
That way a single object can be used to create Requests, Uris and Streams.